chore: move pipeline setup to node-builder
