+++ crates/node-builder/Cargo.toml
@@ -35,6 +35,7 @@ reth-static-file.workspace = true
 reth-prune.workspace = true
 reth-stages.workspace = true
 reth-config.workspace = true
+reth-downloaders.workspace = true
 
 ## async
 futures.workspace = true
