+++ crates/node-builder/src/lib.rs
@@ -26,6 +26,8 @@ pub use handle::NodeHandle;
 
 pub mod rpc;
 
+pub mod setup;
+
 /// Support for installing the ExExs (execution extensions) in a node.
 pub mod exex;
